move galaxy size and ellipticity to morphology
