+++ skypy/galaxy/morphology.py
@@ -1,18 +1,23 @@
-r"""Galaxy size module.
+"""Galaxy morphology module.
 
-This modules computes the angular size of galaxies from their physical size.
+This module provides facilities to sample the sizes and ellipticities of
+galaxies.
 """
-
-import numpy as np
-from astropy import units
-
 
 __all__ = [
     'angular_size',
+    'beta_ellipticity',
     'early_type_lognormal',
     'late_type_lognormal',
     'linear_lognormal',
+    'ryden04',
 ]
+
+
+import numpy as np
+from scipy import stats
+from astropy import units
+from ..utils import random
 
 
 def angular_size(physical_size, redshift, cosmology):
@@ -38,7 +43,7 @@ def angular_size(physical_size, redshift, cosmology):
     Examples
     --------
     >>> from astropy import units
-    >>> from skypy.galaxy.size import angular_size
+    >>> from skypy.galaxy.morphology import angular_size
     >>> from astropy.cosmology import Planck15
     >>> r = angular_size(10*units.kpc, 1, Planck15)
 
@@ -51,6 +56,55 @@ def angular_size(physical_size, redshift, cosmology):
     angular_size = np.arctan(physical_size / distance)
 
     return angular_size
+
+
+def beta_ellipticity(e_ratio, e_sum, size=None):
+    r'''Galaxy ellipticities sampled from a reparameterized beta distribution.
+
+    The ellipticities follow a beta distribution parameterized by
+    :math:`e_{\rm ratio}` and :math:`e_{\rm sum}` as presented in [1]_ Section
+    III.A.
+
+    Parameters
+    ----------
+    e_ratio : array_like
+        Mean ellipticity of the distribution, must be between 0 and 1.
+    e_sum : array_like
+        Parameter controlling the width of the distribution, must be positive.
+
+    Notes
+    -----
+    The probability distribution function :math:`p(e)` for ellipticity
+    :math:`e` is given by a beta distribution:
+
+    .. math::
+
+        p(e) \sim \frac{\Gamma(a+b)}{\Gamma(a) \Gamma(b)} x^{a-1} (1-x)^{b-1}
+
+    for :math:`0 <= e <= 1`, :math:`a = e_{\rm sum} e_{\rm ratio}`,
+    :math:`b = e_{\rm sum} (1 - e_{\rm ratio})`, :math:`0 < e_{\rm ratio} < 1`
+    and :math:`e_{\rm sum} > 0`, where :math:`\Gamma` is the gamma function.
+
+    References
+    ----------
+    .. [1] Kacprzak T., Herbel J., Nicola A. et al., arXiv:1906.01018
+
+    Examples
+    --------
+    Sample 10000 random variates from the Kacprzak model with
+    :math:`e_{\rm ratio} = 0.5` and :math:`e_{\rm sum} = 1.0`:
+
+    >>> from skypy.galaxy.morphology import beta_ellipticity
+    >>> ellipticity = beta_ellipticity(0.5, 1.0, size=10000)
+
+    '''
+
+    # convert to beta distribution parameters
+    a = e_sum * e_ratio
+    b = e_sum * (1.0 - e_ratio)
+
+    # sample from the beta distribution
+    return np.random.beta(a, b, size)
 
 
 def late_type_lognormal(magnitude, alpha, beta, gamma, M0, sigma1, sigma2,
@@ -87,12 +141,12 @@ def late_type_lognormal(magnitude, alpha, beta, gamma, M0, sigma1, sigma2,
     Examples
     --------
     >>> import numpy as np
-    >>> from skypy.galaxy import size
+    >>> from skypy.galaxy import morphology
     >>> magnitude = -16.0
     >>> alpha, beta, gamma, M0 = 0.21, 0.53, -1.31, -20.52
     >>> sigma1, sigma2 = 0.48, 0.25
-    >>> s = size.late_type_lognormal(magnitude, alpha, beta, gamma, M0,\
-                                     sigma1, sigma2)
+    >>> s = morphology.late_type_lognormal(magnitude, alpha, beta, gamma, M0,
+    ...                                    sigma1, sigma2)
 
 
     References
@@ -147,11 +201,11 @@ def early_type_lognormal(magnitude, a, b, M0, sigma1, sigma2, size=None):
     Examples
     --------
     >>> import numpy as np
-    >>> from skypy.galaxy import size
+    >>> from skypy.galaxy import morphology
     >>> magnitude = -20.0
     >>> a, b, M0 = 0.6, -4.63, -20.52
     >>> sigma1, sigma2 = 0.48, 0.25
-    >>> s = size.early_type_lognormal(magnitude, a, b, M0, sigma1, sigma2)
+    >>> s = morphology.early_type_lognormal(magnitude, a, b, M0, sigma1, sigma2)
 
 
     References
@@ -198,10 +252,10 @@ def linear_lognormal(magnitude, a_mu, b_mu, sigma, size=None):
     Examples
     --------
     >>> import numpy as np
-    >>> from skypy.galaxy import size
+    >>> from skypy.galaxy import morphology
     >>> magnitude = -20.0
     >>> a_mu, b_mu, sigma =-0.24, -4.63, 0.4
-    >>> s = size.linear_lognormal(magnitude, a_mu, b_mu, sigma)
+    >>> s = morphology.linear_lognormal(magnitude, a_mu, b_mu, sigma)
 
     References
     ----------
@@ -213,3 +267,68 @@ def linear_lognormal(magnitude, a_mu, b_mu, sigma, size=None):
 
     return late_type_lognormal(magnitude, -a_mu / 0.4, -a_mu / 0.4,
                                b_mu, -np.inf, sigma, sigma, size=size)
+
+
+def ryden04(mu_gamma, sigma_gamma, mu, sigma, size=None):
+    r'''Ellipticity distribution of Ryden (2004).
+
+    The ellipticity is sampled by randomly projecting a 3D ellipsoid with
+    principal axes :math:`A > B > C` [1]_. The distribution of the axis ratio
+    :math:`\gamma = C/A` is a truncated normal with mean :math:`\mu_\gamma` and
+    standard deviation :math:`\sigma_\gamma`. The distribution of
+    :math:`\epsilon = \log(1 - B/A)` is truncated normal with mean :math:`\mu`
+    and standard deviation :math:`\sigma`.
+
+    Parameters
+    ----------
+    mu_gamma : array_like
+        Mean of the truncated Gaussian for :math:`\gamma`.
+    sigma_gamma : array_like
+        Standard deviation for :math:`\gamma`.
+    mu : array_like
+        Mean of the truncated Gaussian for :math:`\epsilon`.
+    sigma : array_like
+        Standard deviation for :math:`\epsilon`.
+    size : int or tuple of ints or None
+        Size of the sample. If `None` the size is inferred from the parameters.
+
+    References
+    ----------
+    .. [1] Ryden B. S., 2004, ApJ, 601, 214
+
+    Examples
+    --------
+    Sample 10000 random variates from the Ryden (2004) model with parameters
+    :math:`\mu_\gamma = 0.222`, :math:`\sigma_\gamma = 0.056`,
+    :math:`\mu = -1.85`, and :math:`\sigma = 0.89`.
+
+    >>> from skypy.galaxy.morphology import ryden04
+    >>> ellipticity = ryden04(0.222, 0.056, -1.85, 0.89, size=10000)
+
+    '''
+
+    # get size if not given
+    if size is None:
+        size = np.broadcast(mu_gamma, sigma_gamma, mu, sigma).shape
+
+    # truncation for gamma standard normal
+    a_gam = np.divide(np.negative(mu_gamma), sigma_gamma)
+    b_gam = np.divide(np.subtract(1, mu_gamma), sigma_gamma)
+
+    # truncation for log(epsilon) standard normal
+    a_eps = -np.inf
+    b_eps = np.divide(np.negative(mu), sigma)
+
+    # draw gamma and epsilon from truncated normal -- eq.s (10)-(11)
+    gam = stats.truncnorm.rvs(a_gam, b_gam, mu_gamma, sigma_gamma, size=size)
+    eps = np.exp(stats.truncnorm.rvs(a_eps, b_eps, mu, sigma, size=size))
+
+    # scipy 1.5.x bug: make scalar if size is empty
+    if size == () and not np.isscalar(gam):  # pragma: no cover
+        gam, eps = gam.item(), eps.item()
+
+    # random projection of random triaxial ellipsoid
+    q = random.triaxial_axis_ratio(1-eps, gam)
+
+    # return the ellipticity
+    return (1-q)/(1+q)
