+++ skypy/galaxy/tests/test_import.py
@@ -1,8 +1,7 @@
 def test_import():
     import skypy.galaxy
-    import skypy.galaxy.ellipticity
     import skypy.galaxy.luminosity
+    import skypy.galaxy.morphology
     import skypy.galaxy.redshift
-    import skypy.galaxy.size
     import skypy.galaxy.spectrum
     import skypy.galaxy.stellar_mass
